Change: Split runtime into separate files
