+++ source/OpenBveApi/Runtime/Route/Station.cs
