+++ source/OpenBveApi/Runtime/System/Time.cs
@@ -1,0 +1,34 @@
+﻿namespace OpenBveApi.Runtime
+{
+	/// <summary>Represents a time.</summary>
+	public class Time
+	{
+		/// <summary>The time in seconds.</summary>
+		private readonly double MyValue;
+
+		/// <summary>Gets the time in seconds.</summary>
+		public double Seconds
+		{
+			get
+			{
+				return this.MyValue;
+			}
+		}
+
+		/// <summary>Gets the time in milliseconds.</summary>
+		public double Milliseconds
+		{
+			get
+			{
+				return 1000.0 * this.MyValue;
+			}
+		}
+
+		/// <summary>Creates a new instance of this class.</summary>
+		/// <param name="value">The time in seconds.</param>
+		public Time(double value)
+		{
+			this.MyValue = value;
+		}
+	}
+}
