+++ source/OpenBveApi/Runtime/Train/HornTypes.cs
@@ -1,0 +1,13 @@
+﻿namespace OpenBveApi.Runtime
+{
+	/// <summary>Represents the type of horn.</summary>
+	public enum HornTypes
+	{
+		/// <summary>The primary horn. The numerical value of this constant is 0.</summary>
+		Primary = 1,
+		/// <summary>The secondary horn. The numerical value of this constant is 1.</summary>
+		Secondary = 2,
+		/// <summary>The music horn. The numerical value of this constant is 2.</summary>
+		Music = 3
+	}
+}
